small fixes, anti hack
I blocked no user spams and found out something about a hacker that is currently interested in the website. He posted links in the comments. When a user clicked on the link, there was a chance for the session to be hijacked. But we have expiring sessions which makes more sense than long sessions in that context. Disabling links in the comments would fix this better.